refactor(account): apply changes required on from code review
