+++ api_account_test.go
@@ -2,8 +2,10 @@ package blockfrost_test
 
 import (
 	"context"
+	"io/ioutil"
 	"net/http"
 	"net/http/httptest"
+	"path/filepath"
 	"testing"
 	"time"
 
@@ -18,20 +20,12 @@ func TestResourceAccount(t *testing.T) {
 	s := httptest.NewServer(
 		http.HandlerFunc(func(w http.ResponseWriter, _ *http.Request) {
 			time.Sleep(20 * time.Millisecond)
-			w.Write([]byte(
-				`{
-					"stake_address": "stake1ux3g2c9dx2nhhehyrezyxpkstartcqmu9hk63qgfkccw5rqttygt7",
-					"active": true,
-					"active_epoch": 412,
-					"controlled_amount": "619154618165",
-					"rewards_sum": "319154618165",
-					"withdrawals_sum": "12125369253",
-					"reserves_sum": "319154618165",
-					"treasury_sum": "12000000",
-					"withdrawable_amount": "319154618165",
-					"pool_id": "pool1pu5jlj4q9w9jlxeu370a3c9myx47md5j5m2str0naunn2q3lkdy"
-					}`,
-			))
+			fp := filepath.Join(testdata, "json", "account", "account.json")
+			file, err := ioutil.ReadFile(fp)
+			if err != nil {
+				t.Fatalf("an error ocurred while trying to read json test file %s", fp)
+			}
+			w.Write(file)
 		}),
 	)
 	defer s.Close()
@@ -60,20 +54,12 @@ func TestResourceAccountRewardsHistory(t *testing.T) {
 	s := httptest.NewServer(
 		http.HandlerFunc(func(w http.ResponseWriter, _ *http.Request) {
 			time.Sleep(20 * time.Millisecond)
-			w.Write([]byte(
-				`[
-					{
-					"epoch": 215,
-					"amount": "12695385",
-					"pool_id": "pool1pu5jlj4q9w9jlxeu370a3c9myx47md5j5m2str0naunn2q3lkdy"
-					},
-					{
-					"epoch": 216,
-					"amount": "3586329",
-					"pool_id": "pool1pu5jlj4q9w9jlxeu370a3c9myx47md5j5m2str0naunn2q3lkdy"
-					}
-					]`,
-			))
+			fp := filepath.Join(testdata, "json", "account", "account_rewards_history.json")
+			file, err := ioutil.ReadFile(fp)
+			if err != nil {
+				t.Fatalf("an error ocurred while trying to read json test file %s", fp)
+			}
+			w.Write(file)
 		}),
 	)
 	defer s.Close()
@@ -106,20 +92,12 @@ func TestResourceAccountHistory(t *testing.T) {
 	s := httptest.NewServer(
 		http.HandlerFunc(func(w http.ResponseWriter, _ *http.Request) {
 			time.Sleep(20 * time.Millisecond)
-			w.Write([]byte(
-				`[
-					{
-					"active_epoch": 210,
-					"amount": "12695385",
-					"pool_id": "pool1pu5jlj4q9w9jlxeu370a3c9myx47md5j5m2str0naunn2q3lkdy"
-					},
-					{
-					"active_epoch": 211,
-					"amount": "22695385",
-					"pool_id": "pool1pu5jlj4q9w9jlxeu370a3c9myx47md5j5m2str0naunn2q3lkdy"
-					}
-					]`,
-			))
+			fp := filepath.Join(testdata, "json", "account", "account_history.json")
+			file, err := ioutil.ReadFile(fp)
+			if err != nil {
+				t.Fatalf("an error ocurred while trying to read json test file %s", fp)
+			}
+			w.Write(file)
 		}),
 	)
 	defer s.Close()
@@ -153,22 +131,12 @@ func TestResourceAccountDelegationHistory(t *testing.T) {
 	s := httptest.NewServer(
 		http.HandlerFunc(func(w http.ResponseWriter, _ *http.Request) {
 			time.Sleep(20 * time.Millisecond)
-			w.Write([]byte(
-				`[
-					{
-					"active_epoch": 210,
-					"tx_hash": "2dd15e0ef6e6a17841cb9541c27724072ce4d4b79b91e58432fbaa32d9572531",
-					"amount": "12695385",
-					"pool_id": "pool1pu5jlj4q9w9jlxeu370a3c9myx47md5j5m2str0naunn2q3lkdy"
-					},
-					{
-					"active_epoch": 242,
-					"tx_hash": "1a0570af966fb355a7160e4f82d5a80b8681b7955f5d44bec0dde628516157f0",
-					"amount": "12691385",
-					"pool_id": "pool1kchver88u3kygsak8wgll7htr8uxn5v35lfrsyy842nkscrzyvj"
-					}
-					]`,
-			))
+			fp := filepath.Join(testdata, "json", "account", "account_delegation_history.json")
+			file, err := ioutil.ReadFile(fp)
+			if err != nil {
+				t.Fatalf("an error ocurred while trying to read json test file %s", fp)
+			}
+			w.Write(file)
 		}),
 	)
 	defer s.Close()
@@ -200,18 +168,12 @@ func TestResourceAccountRegistrationHistory(t *testing.T) {
 	s := httptest.NewServer(
 		http.HandlerFunc(func(w http.ResponseWriter, _ *http.Request) {
 			time.Sleep(20 * time.Millisecond)
-			w.Write([]byte(
-				`[
-					{
-					"tx_hash": "2dd15e0ef6e6a17841cb9541c27724072ce4d4b79b91e58432fbaa32d9572531",
-					"action": "registered"
-					},
-					{
-					"tx_hash": "1a0570af966fb355a7160e4f82d5a80b8681b7955f5d44bec0dde628516157f0",
-					"action": "deregistered"
-					}
-					]`,
-			))
+			fp := filepath.Join(testdata, "json", "account", "account_registration_history.json")
+			file, err := ioutil.ReadFile(fp)
+			if err != nil {
+				t.Fatalf("an error ocurred while trying to read json test file %s", fp)
+			}
+			w.Write(file)
 		}),
 	)
 	defer s.Close()
@@ -243,18 +205,12 @@ func TestResourceAccountWithdrawalHistory(t *testing.T) {
 	s := httptest.NewServer(
 		http.HandlerFunc(func(w http.ResponseWriter, _ *http.Request) {
 			time.Sleep(20 * time.Millisecond)
-			w.Write([]byte(
-				`[
-					{
-					"tx_hash": "48a9625c841eea0dd2bb6cf551eabe6523b7290c9ce34be74eedef2dd8f7ecc5",
-					"amount": "454541212442"
-					},
-					{
-					"tx_hash": "4230b0cbccf6f449f0847d8ad1d634a7a49df60d8c142bb8cc2dbc8ca03d9e34",
-					"amount": "97846969"
-					}
-					]`,
-			))
+			fp := filepath.Join(testdata, "json", "account", "account_withdrawal_history.json")
+			file, err := ioutil.ReadFile(fp)
+			if err != nil {
+				t.Fatalf("an error ocurred while trying to read json test file %s", fp)
+			}
+			w.Write(file)
 		}),
 	)
 	defer s.Close()
@@ -286,18 +242,12 @@ func TestResourceAccountMIRHistory(t *testing.T) {
 	s := httptest.NewServer(
 		http.HandlerFunc(func(w http.ResponseWriter, _ *http.Request) {
 			time.Sleep(20 * time.Millisecond)
-			w.Write([]byte(
-				`[
-					{
-					"tx_hash": "69705bba1d687a816ff5a04ec0c358a1f1ef075ab7f9c6cc2763e792581cec6d",
-					"amount": "2193707473"
-					},
-					{
-					"tx_hash": "baaa77b63d4d7d2bb3ab02c9b85978c2092c336dede7f59e31ad65452d510c13",
-					"amount": "14520198574"
-					}
-					]`,
-			))
+			fp := filepath.Join(testdata, "json", "account", "account_mir_history.json")
+			file, err := ioutil.ReadFile(fp)
+			if err != nil {
+				t.Fatalf("an error ocurred while trying to read json test file %s", fp)
+			}
+			w.Write(file)
 		}),
 	)
 	defer s.Close()
@@ -328,16 +278,12 @@ func TestResourceAccountAssociatedAddresses(t *testing.T) {
 	s := httptest.NewServer(
 		http.HandlerFunc(func(w http.ResponseWriter, _ *http.Request) {
 			time.Sleep(20 * time.Millisecond)
-			w.Write([]byte(
-				`[
-					{
-						"address": "addr1qx2kd28nq8ac5prwg32hhvudlwggpgfp8utlyqxu6wqgz62f79qsdmm5dsknt9ecr5w468r9ey0fxwkdrwh08ly3tu9sy0f4qd"
-					},
-					{
-						"address": "addr1qys3czp8s9thc6u2fqed9yq3h24nyw28uk0m6mkgn9dkckjf79qsdmm5dsknt9ecr5w468r9ey0fxwkdrwh08ly3tu9suth4w4"
-					}
-				]`,
-			))
+			fp := filepath.Join(testdata, "json", "account", "account_associated_addresses.json")
+			file, err := ioutil.ReadFile(fp)
+			if err != nil {
+				t.Fatalf("an error ocurred while trying to read json test file %s", fp)
+			}
+			w.Write(file)
 		}),
 	)
 	defer s.Close()
@@ -369,18 +315,12 @@ func TestResourceAccountAssociatedAssets(t *testing.T) {
 	s := httptest.NewServer(
 		http.HandlerFunc(func(w http.ResponseWriter, _ *http.Request) {
 			time.Sleep(20 * time.Millisecond)
-			w.Write([]byte(
-				`[
-					{
-					"unit": "d5e6bf0500378d4f0da4e8dde6becec7621cd8cbf5cbb9b87013d4cc537061636542756433343132",
-					"quantity": "1"
-					},
-					{
-					"unit": "b0d07d45fe9514f80213f4020e5a61241458be626841cde717cb38a76e7574636f696e",
-					"quantity": "125"
-					}
-					]`,
-			))
+			fp := filepath.Join(testdata, "json", "account", "account_associated_assets.json")
+			file, err := ioutil.ReadFile(fp)
+			if err != nil {
+				t.Fatalf("an error ocurred while trying to read json test file %s", fp)
+			}
+			w.Write(file)
 		}),
 	)
 	defer s.Close()
